models/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# cli\cli.py: 1
# cli\commands\cmd_scripts.py: 1
Click == 7.0

# server\app.py: 1
Flask == 1.1.1

# server\classes\google.py: 9,10
# server\classes\input.py: 5
# server\functions\image.py: 3,4
Pillow == 7.0.0

# server\classes\imgflip.py: 4
beautifulsoup4 == 4.8.2

# server\classes\google.py: 6
google-images-download == 2.8.0

# server\classes\template_clf.py: 7,8
keras == 2.3.1

# server\classes\template_clf.py: 11
matplotlib == 3.1.3

# server\classes\input.py: 2
# server\classes\template_clf.py: 10
# server\functions\misc.py: 2
numpy == 1.18.1

# server\classes\input.py: 6
# server\functions\misc.py: 1
opencv-python == 4.2.0.32

# server\classes\google.py: 2
# server\classes\imgflip.py: 2
pandas == 1.0.1

# config\flask.py: 1
# server\app.py: 2
# server\classes\input.py: 4
# server\sentry.py: 2
python-decouple == 3.3

# server\classes\google.py: 1
# server\classes\imgflip.py: 1
# server\classes\input.py: 1
# server\functions\image.py: 1
requests == 2.22.0

# server\app.py: 4
# server\sentry.py: 1
sentry-sdk == 0.14.1

# setup.py: 1
setuptools == 45.2.0

# not found by pigar
gunicorn == 20.0.4
sentry-sdk[flask] == 0.14.1
lxml == 4.5.0
tensorflow == 2.1.0